requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0

# Web API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter dependencies
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.7.0

# Optional: For enhanced features
redis>=5.0.0  # For caching (future)
prometheus-client>=0.17.0  # For metrics (future)
